[refactor] Move text comparison helpers to their own module
