+++ src/_pytest/assertion/util.py
@@ -22,8 +22,10 @@ from _pytest.assertion._compare_sequence import _compare_eq_iterable
 from _pytest.assertion._compare_sequence import _compare_eq_sequence
 from _pytest.assertion._compare_set import _compare_eq_set
 from _pytest.assertion._compare_set import SET_COMPARISON_FUNCTIONS
+from _pytest.assertion._compare_text import _compare_eq_text
+from _pytest.assertion._compare_text import _notin_text
+from _pytest.assertion._typing import _AssertionTextDiffStyle
 from _pytest.assertion._typing import _HighlightFunc
-from _pytest.compat import assert_never
 from _pytest.config import Config
 from _pytest.config import UsageError
 
@@ -42,7 +44,6 @@ _assertion_pass: Callable[[int, str, str], None] | None = None
 _config: Config | None = None
 
 ASSERTION_TEXT_DIFF_STYLE_INI = "assertion_text_diff_style"
-_AssertionTextDiffStyle = Literal["ndiff", "block"]
 ASSERTION_TEXT_DIFF_STYLE_NDIFF: Literal["ndiff"] = "ndiff"
 ASSERTION_TEXT_DIFF_STYLE_BLOCK: Literal["block"] = "block"
 ASSERTION_TEXT_DIFF_STYLE_CHOICES = (
@@ -319,91 +320,6 @@ def _compare_eq_any(
     return explanation
 
 
-def _compare_eq_text(
-    left: str,
-    right: str,
-    highlighter: _HighlightFunc,
-    verbose: int,
-    assertion_text_diff_style: _AssertionTextDiffStyle,
-) -> list[str]:
-    match assertion_text_diff_style:
-        case "block":
-            return _diff_text_block(left, right)
-        case "ndiff":
-            return _diff_text(left, right, highlighter, verbose)
-        case unreachable:
-            assert_never(unreachable)
-
-
-def _diff_text_block(left: str, right: str) -> list[str]:
-    return [
-        "Left:",
-        *_format_text_block_lines(left),
-        "",
-        "Right:",
-        *_format_text_block_lines(right),
-    ]
-
-
-def _format_text_block_lines(text: str) -> list[str]:
-    return [f"  {line}" for line in text.split("\n")]
-
-
-def _diff_text(
-    left: str, right: str, highlighter: _HighlightFunc, verbose: int = 0
-) -> list[str]:
-    """Return the explanation for the diff between text.
-
-    Unless --verbose is used this will skip leading and trailing
-    characters which are identical to keep the diff minimal.
-    """
-    from difflib import ndiff
-
-    explanation: list[str] = []
-
-    if verbose < 1:
-        i = 0  # just in case left or right has zero length
-        for i in range(min(len(left), len(right))):
-            if left[i] != right[i]:
-                break
-        if i > 42:
-            i -= 10  # Provide some context
-            explanation = [
-                f"Skipping {i} identical leading characters in diff, use -v to show"
-            ]
-            left = left[i:]
-            right = right[i:]
-        if len(left) == len(right):
-            for i in range(len(left)):
-                if left[-i] != right[-i]:
-                    break
-            if i > 42:
-                i -= 10  # Provide some context
-                explanation += [
-                    f"Skipping {i} identical trailing "
-                    "characters in diff, use -v to show"
-                ]
-                left = left[:-i]
-                right = right[:-i]
-    keepends = True
-    if left.isspace() or right.isspace():
-        left = repr(str(left))
-        right = repr(str(right))
-        explanation += ["Strings contain only whitespace, escaping them using repr()"]
-    # "right" is the expected base against which we compare "left",
-    # see https://github.com/pytest-dev/pytest/issues/3333
-    explanation.extend(
-        highlighter(
-            "\n".join(
-                line.strip("\n")
-                for line in ndiff(right.splitlines(keepends), left.splitlines(keepends))
-            ),
-            lexer="diff",
-        ).splitlines()
-    )
-    return explanation
-
-
 def _compare_eq_cls(
     left: object,
     right: object,
@@ -463,22 +379,3 @@ def _compare_eq_cls(
                 )
             ]
     return explanation
-
-
-def _notin_text(term: str, text: str, verbose: int = 0) -> list[str]:
-    index = text.find(term)
-    head = text[:index]
-    tail = text[index + len(term) :]
-    correct_text = head + tail
-    diff = _diff_text(text, correct_text, dummy_highlighter, verbose)
-    newdiff = [f"{saferepr(term, maxsize=42)} is contained here:"]
-    for line in diff:
-        if line.startswith("Skipping"):
-            continue
-        if line.startswith("- "):
-            continue
-        if line.startswith("+ "):
-            newdiff.append("  " + line[2:])
-        else:
-            newdiff.append(line)
-    return newdiff
